Stablize BTDiscoveryTest
We used to set a new name for the BT target device but that new name sometimes is not ready fast enough for discoverer to see.
This flakiness doesn't cause too much issue for postsubmit since it usually passes on the retry.
However, when running BTDiscoveryTest as presubmit, there is no retry and the test is too unstable.

Bug: 384052783
Test: https://paste.googleplex.com/6352981555150848

Change-Id: I436710c9dcdcc2cf7b48ceed80d8402fef5c6c90