[shade] check for battery composable if flags are enabled
With the new_status_bar_icons flag, the view structure of the shade is
changed now. The battery composable contains the percent text, so we
shouldn't be looking for that view at all.

This change will check for the battery composable as defined in
systemui's ids.xml.

Test: QS platform scenario tests
Bug: 398226765
Flag: EXEMPT test fix
Change-Id: If1391bf7d8600bec1d6f7757a966e3d40f36c49a