VKMS: Add a tradefed (atest) wrapper using VKMS.
Adding a new mechanism to exercise VKMS using any existing tradefed test
(atest).

[Why]
Now that we have VKMS, we can create any emulate the hardware resouces
on the AVD. Using this, we can configure the hardware in a way we want
before running a test such as the VTS tests targeting the HWC.
With the wrapper, at setup, the hardware emulation will be configured,
then the test will run against these configuration, eliminating the need
to perform any manual configuration before the test runs, or the need
for hardware if it can be emulated using VKMS.
The test also runs like any other `atest` with the bonus of having a
config.

[How]
Create a setup_vkms_connectors_for_atest that configures VKMS through
AndroidTest.xml
Create teardown_vkms that cleans up VKMS after the test is done.
Create cc_defaults to take care of the setup
Create 2 useful example tests to showcase how to use it

[To create your own test]
1. Create a new directory under `atest_wrapper`.
2. Create an Android.bp that that builds an existing atest which we want
   to run
3. Create an AndroidTest.xml which calls
   `setup_vkms_connectors_for_atest` that configure vkms on test setup.

Bug: b/402431306, b/352790467
Test: atest Vkms5WithConfig_VtsHalGraphicsComposer3 and atest Vkms2VtsHalGraphicsComposer3
Flag: TEST_ONLY
Change-Id: I6641a0846769e6eca1561913f6c925a3bde59eef