Remove PerfSetupTests
This test was used as a helper for running perf-setup.sh. However, that
script is now installed by default on eng and userdebug builds, so the
PerfSetupTests test was no longer helpful. It was also broken since
as it was currently configured, it could not find perf-setup.sh in its
test artifacts. We could have updated it to run 'adb shell
perf-setup.sh', and left it as an example, but usage of perf-setup.sh
should be self-explanatory.

Test: atest PerfSetupTests
Bug: 169272096
Change-Id: Ia8575fc25ad86c8058d8bcb9901f5c9ac633f5e2