[CD Testing] Add DesktopMouseTestRule utils to traverse displays
Please refer to this doc for more information and demo
http://shortn/_srD9nWXkBB

The main idea here is to do graph traversal to find list of displays as
the path, then at every display, move mouse to the midpoint
of intersection of 2 displays. Then move slightly to the next display
to cross display.

Will add the move() in the next CL

Bug: 399523818
Test: presubmit
Flag: EXEMPT developing test rule

Change-Id: I806d0408da9a444de916c50c0c910d356e3df6fa