Allow to disable metric collection via a dedicated arg
Adding `disable` arg which is equivalent to `log` arg to completely
disable metric collection in the test runner.

This is needed as having "log" with value "true" in many androidTest.xml confused countless developers.

A possible solution would be to remove the collector completely when not needed, but providing the arg allows for easier override from cli and abtd ui.

Bug: 410773677
Flag: NONE - test code
Test: Sucessful test run of sysui tests
Change-Id: Ia9cdfa332c5f5f6900510856d9c986dd579fdaf5