read flag value through public api
This change changes the behavior of DeviceFlagsValueProvider

* if the flag is read_write, and if the test is running on B+ devices,
  then the provider will read the flag value from the new storage.
* if the flag is read_only or the test is running on devices before B,
  then it will still use reflection to read the flag value.

The reason for this change is in the generated flag accessor code, it
will compare the flag fingerprint, which is created at build time. If
the fingerpint doesn't match  then the flag will return the default
value, which is captured at the build time. However this is not the
ideal behavior for this tool. If this test is not running on a release
build, then this test tool expects to read the flag value from the
device.

Test: atest FlagJUnitTests FlagJUnitHostTests
Bug: 412743688
Change-Id: I35e4d0934dad5c038997cf3f33c8d59e45573901