Add check for dismiss view being shown for bubbles
Bubbles drag to dismiss interaction is reported as flaky.
The existing impl did not check if the dismiss view is shown or not, and
just dragged the bubble to the bottom of the screen.
Update the drag logic to check where the dismiss view is and finish the
drag in the center of the dismiss view.
This should ensure that the bubble is always dragged into the dismiss
view.

Bug: 411685056
Test: atest PlatformScenarioTests:android.platform.test.scenario.sysui.bubble.DragAndDismiss
Flag: EXEMPT, updating test
Change-Id: I0adcc34f28e3155a57ccc37f4069344ab002aa18