Do not look for bubble bar after collapsing it
When bubble bar is collapsed, it may be stashed, if we are in an app.
The previous implementation for collapsing the expanded view always
attempted to look for the bubble bar object.
But it may not always exist. Remove the step for looking for bubble bar
so this method can be used when bubble bar is shown on top of an app or
on home screen.

Bug: 357115534
Test: atest PlatformScenarioTests:BubbleBarTest
Flag: EXEMPT, updating test
Change-Id: I56d8ddc441a6c18350804277279c74e8372b41f1