Create a more robust API to expand bubble bar
When bubble bar is shown on top of an app, it can stash back to the
handle. Create an API that can expand the bubble bar in case it is
currently stashed to a handle or it is currently shown on screen.
To avoid a long wait in case the bar or handle is already visible, it
will first check if these views are present. If they are, we will try to
click on the immediately. If they are not visible, we will wait for
either of them appear.
In case bubble bar is currently visible, click on it can still fail as
the bar is only shown transiently. And by the time we try to click it,
it already may have been animated back to handle. Ignore the failure
from clicking on the bar initially.

Bug: 415230125
Test: atest PlatformScenarioTests:BubbleBarTest
Flag: EXEMPT, test
Change-Id: Id7ecf0a91d15e3b1943cc163387409d44116f843