Initialize lastTime when swipe starts
Initialize lastTime when down event is sent.
lastTime is used to calculate the eventTime that is sent with the mock
motion event.
Previously lastTime was not properly initialized, which resulted in
motion events where event downTime was larger than eventTime.
This broke cases where motion event handlers check for event time
compared to down time. One example is DragDetector that is used to
determine if motion events originating from app handle are a drag or
not.

Bug: 415111575
Flag: EXEMPT, updating test
Test: atest PlatformScenarioTests:BubbleAnythingTest
Change-Id: If183e78d803f568d23a3a7dfbfecbbcb4b294435