Remove troublesome injectInputEvent parameter
Transferring latest changes into Google3 resulted in error because of calling injectInputEvent with 3 parameters. This overloaded version of the method is marked as @TestApi and hidden, making it not recognizable on the G3 side of things. This CL removes the troublesome parameter and uses method recognized in both repos, based on a conversation with @nicomazz.

Bug: b/404797238
Change-Id: I36eff5a6ce06ac231db92d0b3dfb52f69ef16d34