Make dual target tests more robust
When a dual target tile is small, clicking on it toggles it. Make the
tests more robust by performing the appropriate action to open the
dialog based on the size of the tile.

As of now, this is not an issue because all dual target tiles start
large, but this will help when we have dual target tiles that start
small (like flashlight).

Also, introduce a way to obtain the toggleable behavior from a dual
target tile, regardless of its current size.

Test: atest android.platform.test.scenario.sysui.quicksettings.composetiles.qs
Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Fixes: 418209295
Change-Id: I0892fbb16143065627387aa9f1c9e387688e136f