Add hasNoActivityOnDisplay() to WindowManagerState
This new API will be used in e2e tests for checking that there is no
activity on a mirroring display.

This CL also makes some small changes to the Javadoc of some APIs in
ConditionsFactory to ensure consistency with others in the same class.

Bug: 392787946
Test: EnableDisplayMirroringSwitchTest
Flag: EXEMPT adding utility methods
Change-Id: Ib2ac52d705d92a53f3e213f495d923d1dd4ddb4f