Create and use btEnableWithLongerWait in AAOS BT tests
AAOS BT tests used to call btEnable from mobly-bundled-snippet which waits a maximum of 2 seconds for the bluetooth permission request dialog to show up. When dialog failed to show within 2 seconds and the test proceeded to click on ALLOW button, btEnable throw NPE.
Since the 2 seconds work for most users of mobly-bundled-snippet and OEM partners, it's suggested to create custom snippet with longer wait for AAOS needs.

Bug: 384052783
Test: https://paste.googleplex.com/6754552575164416

Change-Id: Id928c7b78cdea84875cee14a06a26b26f5864b94