[Flexiglass] Remove SplitShade specific checks from the NotificationShade tapl
This is a low-hanging fruit to keep Notification tests running in DualShade.

The NotificationShade.kt tapl helper was always doing an assertion about
the QS-Notif relation when the config_use_split_notification_shade was
enabled. This assertion was breaking with DualShade, plus it is not
really the right place to do it. Tests that are not dependent on the
shade configuration (e.g. opening a notification) should not fail if
we cannot find quick settings in the shade.

Bug: 419192993
Flag: com.android.systemui.scene_container
Test: atest PlatformScenarioTests:android.platform.test.scenario.sysui.notification
Change-Id: I23f81340d1c7d850cf1fcd909e3b900bd7d3d0c3