Open QS with an adb command
Swiping to open QS adds flakiness risk, especially on virtualized
devices with poor performance. This change adds an option that uses
'adb shell cmd statusbar expand-settings', which I confirmed works
everywhere for all forms of the shade. If it fails for some reason,
it will fall back to the previous method (i.e. swiping). All calls
to the previous method that used swipe have been audited and replaced
with an explicit request to swipe in tests that intend to verify the
swiping behavior.

Bug: 423908119
Test: Ran a sampling of affected tests
Flag: EXEMPT test only change
Change-Id: I10737434e2be867f48ccd7b2376a6f68e4652bfa