Fix NotificationRemoteInput test failures on unlocking cases.
Some NotificationRemoteInput tests failed on unlocking cases. The reason
is after clicking the reply button on the locked shade. Have to do
unlock to see the notification's reply area. The `replyWithText` method
doesn't consider unlock cases that fail when checking for reply
button. So, I added a check to skip the check if
the reply area exists.

Test: https://android-build.corp.google.com/builds/abtd/run/L81200030013252540
Bug: 425223949
Change-Id: Icdf2b51643615bce323825896f737c2852b33328