Optimize BaseOrientationRule
OrientationRule is added to all the tests as its present in
sysUITestBase, hence optimizing this rule would benefit reducing run
time for all tests directly or indirectly applying this rule. Currently
this rule takes 5secs. The goal of this rule is to set the device in
required orientation when the test is annotated with @portrait or
@landscape, and set it in natural orientation for the device (phone -
portrait, tablets - landscape) when no specific orientation is
requested.

This change is a small optimization by caching the current orientation.
Before change, device.stableOrientation(), which returns the current
device orientation, is called thrice back to back. And stableOrientation
call has a wait for 1 sec for the orientation to settle. hence saving
about 2secs per each test run across all tests.

Test: Ran the presubmit suite

Bug: 419192390

Change-Id: Ice72cfeed7c46b5be90e31962793ad7eb2e9e59a