igt: Run IGT with VKMS enabled if on CF
[Why]
IGT is super useful if it runs on Ika with VKMS enabled.
This makes it test part of the drm kernel that ships on real hardware.
On top of that, the configuration allows it to exercise the tests with
simulated connected displays, which exercises a larger part of the code
and test paths.

[How]
In the config template, run a command that checks if the product name
contains "cf_", which would detect that it is a Cuttlefish device.
If so, setup VKMS using the atest setup binary, creating 1 edp, 1 DP and
1 HDMI.
on teardown, do the same check to teardown vkms as well.

Do the changes in the config template instead of a VKMS wrapper means
that any test added to the Android.bp list of IGT tests will
automatically be capable of running on CF with VKMS.
This way, we don't need to duplicate for a VKMS wrapper. The same test
module can be used anywhere, the test setup will be smart enough to
setup the env if it's CF, or do nothing if it's on real HW.

Bug: b/427730909, b/371577699
Test: atest IgtKmsAddfbBasicTestCases & check that `modetest -M vkms`
shows the 3 displays we've setup

Change-Id: I04077d513454b409585c01d374016c2f726238c1