Fix battery saver dialog creation on external display
The battery saver QS tile was using the application context to launch the dialog instead of the shade one. This resulted in the dialog opening in the wrong display.

Now dialogs are opened on the focused display instead (which will be the one where the shade is open in most cases)

Bug: 362719719
Bug: 383294128
Flag: com.android.systemui.shade_window_goes_around
Test: OpenBatterySaverDialogOnOtherDisplay
Change-Id: I197a9dde0145ba0d75bcdc4743a9facbdf54c345