Avoid an extra waitForOrientationToSettle at the end of test
This `waitForOrientationToSettle` seems not necessary and it adds extra time in the test.
Why this is not required - Currently this waits for a fixed 1 sec after resetting the rotation settings to natural.This is not required in the rule clean up, as its performed in the @before of the test to ensure the device is stable and tests are idempotent.

Test: Ran the presubmit suite

Bug: 419192390

Change-Id: I068bd107590a9dd41ef16cd71084650ce8a85424