Allow streaming configuration file to stdin in PerfettoTraceListener
Adds functionality to stream a configuration file directly to perfetto
stdin.

This avoids having to push the configuration file to a specific
location, which tests don't have write access to by default.

This will allow our UiTraceListener to generate a custom config file on
demand and stream it to the perfetto command since the perfetto process
will not have access to the file created by the test app.

Test: n/a
Bug: 433202739
Change-Id: Id45f141a0142a6a1ae7644619c31b9dc33e6857e