IGT: Remove all direct uses of arch subdir
[Why]
IGT build its binaries in a subdirectory with the architecture name such
as `x86_64` which was fine, until you want to start using it on arm.

[How]
Before running the tests, flatten the test directory on the DUT so the
test is accessible without a subdir, directly from `igt_tests/`
As it's flattend, remove the arch name from the igt binary dir.

Bug: b/428237472
Test: atest IgtKmsConcurrentTestCases -c
Flag: Test only
Change-Id: I26741e59bec8a32468122c146de2a7263a1fad30