Ensure correct trace config is used
The UiTraceListener was initializing the tracing strategies from the parent PerfettoListener before adding its own custom trace configuration to the arguments bundle.

This resulted in the PerfettoTracingStrategy instances not being aware of the custom trace config file, leading to traces being collected with the wrong settings or not at all.

By moving the call to super.setupAdditionalArgs() to the end of the method, we ensure that all custom arguments are in the bundle before the tracing strategies are configured.

Change-Id: Ie642a0cc0512b04a9b852769cc56488ce0d61716