Allow passing an effectContext for the composeRule being created.
This allows users of the `ComposeScreenshotTestRule` to pass in the
CoroutineContext used to run the composition. The context for
LaunchedEffects and rememberCoroutineScope is derived from this context.

This allows clients to ensure the effects are dispatched on the correct
thread, for example STL that requires the effects to be run on the same
thread.

Test: Test only
Flag: TEST_ONLY
Change-Id: I2dd639c43324c3a9b3440e510c68d756760392ef