Remove doh_unit_test
The doh and dns_proxy crates are being combined into one and so are the
tests. It is unclear to me how this file is maintained and how new tests
are added to it (as I would presume no one is aware of this file's
existence), so I will abstain from adding the newly created test for now.

If it is maintained manually (doubtful?), then I can send a
follow up.

Test: TH
Flag: EXEMPT mainline
Change-Id: I7135cae54db88377f483b9b56cb8464b27910a75